+++ source/projects/pktpy/tests/test_cpp/test.py
@@ -1,0 +1,3 @@
+def test():
+    return 'hello world'
+