+++ CHANGELOG.md
@@ -17,6 +17,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 ## [0.1.x]
 
+- Added `execfile` method to `pktpy`.
+
 - Added list support to `pktpy`.
 
 - Added examples of wrapping max api and local functions to `pktpy`.